dlete locilizer (unused)
