--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/Placeholder.java
@@ -1,13 +1,0 @@
-package org.firstinspires.ftc.teamcode.subsystems;
-
-import com.acmerobotics.roadrunner.Time;
-import com.acmerobotics.roadrunner.Twist2dDual;
-
-import org.firstinspires.ftc.teamcode.roadrunner.Localizer;
-
-public class Placeholder implements Localizer {
-    @Override
-    public Twist2dDual<Time> update() {
-        return null;
-    }
-}
